refactor: split semantic artifact budgeting
